Add the ability to open TASes by drag and dropping
How this works: every frame lua side checks for the existence of the file
called "file_pending" in the root directory. When it finds it, it knows
that a file has been dropped and loads it in.

"file_pending" is created from the JS side, as well as writing the
actual TAS file dropped to the in-memory filesystem. It knows which file
to write to because the lua side creates and updates a file called
"working_file" which contains exactly the filename that you'd expect it
to contain.

Currently only works with cctas.

Also I accidentally changed whitespace to tabs on random lines. No need
to thank me.