Switch tas:advance_keystate to use keybindings.lua
love.keyreleased became one hack longer. This is related to the fact
that when a one key is released, all others are also considered
released. This behaviour makes sense for keybindings consisting of
several chords, but specifically messes up the thing tas tool is doing
here. I'm thinking of a better solution to this.