Fix numberlua; Remove hack from cart.lua
For some reason web keeps having unique floating point errors that are
not encountered locally. In fact this was the reason numberlua crashed
originally. For authenticity's sake I restored the original bxor
implementation with flooring added. Considering the values a, am, b, bm
are used as indices I believe rounding them to a whole number is safe.

I can't say the same about rounding the output of band and tobit, so an
Epsilon value is used (derived experimentally).

Nowhere does the tastool code try to apply bitops to non-whole values,
so even floor(x+0.5) like in bxor would've been safe, but I'm trying to
look out for custom carts.